Add useIsMounted
