+++ client/hooks/useIsMounted.js
@@ -1,0 +1,15 @@
+import { useEffect, useRef } from 'react';
+
+export const useIsMounted = () => {
+	const mounted = useRef(true);
+
+	useEffect(() => {
+		mounted.current = true;
+
+		return () => {
+			mounted.current = false;
+		};
+	}, [mounted]);
+
+	return () => mounted.current;
+};
